% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PWLCompressor.R
\name{PWLCompressor}
\alias{PWLCompressor}
\title{Compress an edf to a pwl distribution}
\usage{
PWLCompressor(sample, accuracy = 0.01)
}
\arguments{
\item{sample}{A vector with the sample values}

\item{accuracy}{Relative accuracy of the approximation}
}
\value{
A list where X and Y cointains the abscissas and ordinates of the points for the admissible pwl approximation
}
\description{
Compress an edf to a pwl distribution
}
\examples{
##Minimal example
Sample <- c(1,1.6,4.3,4.6,6,7.1,13,13.4,16,18.8)
PWLapprox <- PWLCompressor(Sample, 0.01)
}
\references{
Arbenz, W. and Guevara W. (2015) Risk Measure Preserving Approximation of Univariate Monte Carlo Simulations Results with Insurance Applications \url{https://sites.google.com/site/philipparbenz/home/pwl-approximation}
}

