% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLCompressor.R
\name{PWLCompressor}
\alias{PWLCompressor}
\title{Compress an edf to a pwl distribution}
\usage{
PWLCompressor(sample, accuracy = 0.01, quantile = c(0, 1), minatom = 1000,
  relatom = 0.005, fullcheck = F)
}
\arguments{
\item{sample}{A vector with the sample values}

\item{accuracy}{Relative accuracy of the approximation}

\item{quantile}{A vector with the quantiles which shall be contained in the basis of the approximation                         approximation.}

\item{minatom}{Minimum sample size which needs to be satisfied such that the algorithm looks for atoms}

\item{relatom}{Relative threshold for atoms to be detected}

\item{fullcheck}{A boolean to determine whether the algorithm should check whether that an admissible approximation is also strictly admissible}
}
\value{
A list where X and Y cointains the abscissas and ordinates of the points for the admissible pwl approximation
}
\description{
Compress an edf to a pwl distribution
}
\examples{
##Minimal example
Sample <- c(1,1.6,4.3,4.6,6,7.1,13,13.4,16,18.8)
PWLapprox <- PWLCompressor(Sample, 0.01)
}

